MAINT: Move ErrorSnapshot to exceptions.py
